fix: docker-compose DB config
Setting custom values for Postgres user credentials did not work if the
application running inside the container because by default the Postgres
container creates only a default user with default credentials.
Previously these custom credentials specified in the env variables were
passed to the Postgres image and it led to an error, because this user
didn't exist in the default configuration, so, containers crashed in
this case.

Now if the app runs in containers, user credentials specified in the env
variables are ignored and the default user credentials are used.