fix: remove redundant param from endpoint
The `v1/modules/:module_id/keys` endpoint can accept only two query
params (`used` and `operatorIndex`) that it uses to filter keys from the
output. But previously public interface of this endpoint incorrectly
stated that it also can accept the `moduleAddresses` param. It doesn't
make sense, because module ID is specified in the `:module_id` URL param
for this endpoint.

Now the `moduleAddresses` search param has been removed from the public
interface of the endpoint.