chore: add default values for DB-related envs
Add default values to the database-related envs to the `docker-compose`
file for e2e tests.

These variables are needed for the test process and previously they were
gotten from the `.env` file. Now they are hard-coded in the
docker-compose file, so, an incorrect DB configuration variables in the
`.env` file should not affect the successful run of e2e tests.