feat: sanity checker for network config
Implement a sanity checker that checks that the information stored in
the DB corresponds the chain for which the app is run. If the app is run
for one chain, and the DB contains data for another chain, the sanity
checker prevents the app from start and throws the error.