chore: read password from file in tests
Modify tests that check how the `configService.get` method works if it
needs to read the password from the file provided in the value of the
`DB_PASSWORD_FILE` env variable. Now these tests create a temporary file
with a password before testing the desired behavior (so that the
`configService.get` method can read the password from the real file),
and then remove this temporary file when the test is complete.